requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
PyYAML==6.0.1
python-dotenv==1.0.0
requests==2.31.0 